Warn when app is not using the latest Ruby version
Developers should be using the latest Ruby version for their minor series. For example currently developers
using 2.4.x should be using 2.4.7, if not then this PR will give them a warning:

```
There is a more recent Ruby version available for you to use:

ruby-2.4.7

The latest version will include security and bug fixes, we always recommend
running the latest version of your minor release.

See https://devcenter.heroku.com/articles/ruby-versions for all available versions.
```

Since it is relatively expensive to check each version serially, we can instead queue up 5 version checks at a
time using threads. If the last version check shows that it is a valid version, then it will enqueue another
round of 5 checks in parallel.

We use similar logic for detecting EOL versions. For example when someone is using 2.2 then we traverse the s3
repository, 2.3.0, 2.4.0, 2.5.0, 2.6.0, 2.7.0, etc. In this case there
are more than 3 more recent versions so 2.2 is for sure EOL and we
recommend that people update to the 3rd most recent version in the
above case it would be 2.5.0.

We also have to account for the case that there is a major version bump,
so in addition to the minor version check we will check to see if 3.0.0
is released. If it has been then we will also check for 3.1.0, 3.2.0,
3.3.0, etc.

There is a problem with relying on deployed Ruby versions on S3 to
indicate EOL support since pre-released binaries are not marked. For
example 2.7.0 is currently in "preview" mode, however the binary is at
2.7.0. So when someone using 2.4.x ruby executes the EOL check it will
see that 2.5, 2.6, and 2.7 are out. Because of this we have 2 different
EOL warnings, one for "maybe" eol that warns the version they are on
either is already EOL or is about to become EOL at the next release.

For example in the future when 2.7.0 is released and someone is using
2.5.0 then they will start getting a warning when the 2.8.0 (or 3.0.0)
preview version is released even though 2.5.0 would still technically be
supported until the full release at Christmas. This is reasonable as it
gives people a heads up that they need to upgrade soon. They would keep
getting this "maybe" warning until the preview of 2.9.0 (or 3.1.0) would
be released. At that time they would start to receive the more severe
EOL message.